Update Pyramids-scroller.user.js
For now this works well either for Pyramids or for Solitaire. At least for me because I'm not a fan of Solitaire getting scrolled that low, but if you are, then that's great! Tbh I wonder if others have the same preference as I do. I'm leaving the default as what it was before, i.e. scrolling to the game title bar.